+++ schema.ts
@@ -1546,7 +1546,7 @@ export interface ObjectSchemaDefinition<K extends string, L extends string>
   index?: IndexDefinition;
 
   /**
-   * The name of the property within {@link ObjectSchemaDefinition.properties} that can be be interpreted as
+   * The name of the property within {@link ObjectSchemaDefinition.properties} that can be interpreted as
    * text representing the version of an item.
    *
    * Must be a {@link ValueType.String} property.
@@ -2493,7 +2493,7 @@ export function makeReferenceSchemaFromObjectSchema(
 export function withIdentity(schema: GenericObjectSchema, identityName: string): GenericObjectSchema {
   return {
     ...deepCopy(schema),
-    // Our typing throughout the SDK is dishonest about identities, we declare the the output of
+    // Our typing throughout the SDK is dishonest about identities, we declare the output of
     // formula/table definitions are Identity but they don't actually include `packId` unless the
     // developer specified it, we don't inject the current packId until upload time (and even then
     // I think it only gets injected in the Metadata, not the actual runtime code) and we never really